quickvisa-api/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
selenium
requests
apscheduler
passlib